requirements.txt
opencv-contrib-python>=4.9.0.80
playwright>=1.40.0
Jinja2>=3.1.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
google-genai>=0.7.0
Pillow>=10.2.0
python-dotenv>=1.0.1
slack_bolt>=1.18.0
slack_sdk>=3.26.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
